requirements.txt
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-openai==0.0.6
langsmith==0.1.17

openai==1.25.0

streamlit==1.35.0
streamlit-folium==0.25.0
folium==0.16.0

geopy==2.4.1
requests==2.31.0
python-dotenv==1.0.1
